+++ src/beaminteraction/4C_beaminteraction_beam_to_solid_volume_meshtying_pair_base.cpp
@@ -172,12 +172,12 @@ void BEAMINTERACTION::BeamToSolidVolumeMeshtyingPairBase<Beam, Solid>::get_pair_
     std::vector<double>& point_coordinates = visualization_data.get_point_coordinates();
     std::vector<double>& displacement = visualization_data.get_point_data<double>("displacement");
 
-    std::vector<double>* pair_beam_id = nullptr;
-    std::vector<double>* pair_solid_id = nullptr;
+    std::vector<int>* pair_beam_id = nullptr;
+    std::vector<int>* pair_solid_id = nullptr;
     if (write_unique_ids)
     {
-      pair_beam_id = &(visualization_data.get_point_data<double>("uid_0_pair_beam_id"));
-      pair_solid_id = &(visualization_data.get_point_data<double>("uid_1_pair_solid_id"));
+      pair_beam_id = &(visualization_data.get_point_data<int>("uid_0_pair_beam_id"));
+      pair_solid_id = &(visualization_data.get_point_data<int>("uid_1_pair_solid_id"));
     }
 
     // Loop over the segments on the beam.
@@ -221,12 +221,12 @@ void BEAMINTERACTION::BeamToSolidVolumeMeshtyingPairBase<Beam, Solid>::get_pair_
     std::vector<double>& displacement = visualization_data.get_point_data<double>("displacement");
     std::vector<double>& force = visualization_data.get_point_data<double>("force");
 
-    std::vector<double>* pair_beam_id = nullptr;
-    std::vector<double>* pair_solid_id = nullptr;
+    std::vector<int>* pair_beam_id = nullptr;
+    std::vector<int>* pair_solid_id = nullptr;
     if (write_unique_ids)
     {
-      pair_beam_id = &(visualization_data.get_point_data<double>("uid_0_pair_beam_id"));
-      pair_solid_id = &(visualization_data.get_point_data<double>("uid_1_pair_solid_id"));
+      pair_beam_id = &(visualization_data.get_point_data<int>("uid_0_pair_beam_id"));
+      pair_solid_id = &(visualization_data.get_point_data<int>("uid_1_pair_solid_id"));
     }
 
     // Loop over the segments on the beam.
